Move TokenGenerator out of todo/
Needs follow up in FF. We need to add a copy of the things
to the Membership BC that is presently still there
