+++ src/Authorization/RandomTokenGenerator.php
@@ -2,7 +2,7 @@
 
 declare( strict_types = 1 );
 
-namespace WMDE\Fundraising\Frontend\Infrastructure;
+namespace WMDE\Fundraising\Frontend\DonationContext\Authorization;
 
 /**
  * @licence GNU GPL v2+
