+++ tests/Unit/Authorization/RandomTokenGeneratorTest.php
@@ -2,12 +2,12 @@
 
 declare( strict_types = 1 );
 
-namespace WMDE\Fundraising\Frontend\Tests\Unit\Infrastructure;
+namespace WMDE\Fundraising\Frontend\DonationContext\Tests\Unit\Authorization;
 
-use WMDE\Fundraising\Frontend\Infrastructure\RandomTokenGenerator;
+use WMDE\Fundraising\Frontend\DonationContext\Authorization\RandomTokenGenerator;
 
 /**
- * @covers \WMDE\Fundraising\Frontend\Infrastructure\RandomTokenGenerator
+ * @covers \WMDE\Fundraising\Frontend\DonationContext\Authorization\RandomTokenGenerator
  *
  * @licence GNU GPL v2+
  * @author Jeroen De Dauw < jeroendedauw@gmail.com >
